backend/requirements.txt
flask
requests
pycti
logger
python-dotenv
bs4
flask_cors
pefile
yara-python
networkx